adding other objects to the cpp98 interface

+++ src/helics/cpp98/MessageFederate.hpp
@@ -96,12 +96,12 @@ class MessageFederate : public virtual Federate
     /** Methods for sending a message **/
     void sendMessage (helics_endpoint source, const std::string &dest, const char *data, size_t len)
     {
-        helicsEndpointSendMessageRaw (source, dest.c_str(), data, len);
+        helicsEndpointSendMessageRaw (source, dest.c_str(), data, static_cast<int>(len));
     }
 
     void sendMessage (helics_endpoint source, const std::string &dest, const char *data, size_t len, helics_time_t time)
     {
-        helicsEndpointSendEventRaw (source, dest.c_str(), data, len, time);
+        helicsEndpointSendEventRaw (source, dest.c_str(), data, static_cast<int>(len), time);
     }
 
     void sendMessage (helics_endpoint source, message_t &message)
